Fixing up inf scripts to cause them to stay on the ground for at least 5 seconds on first impact (removes some of the randomness from inf combat).
Added a gadget for inf supply - when in supply range, they get a firing bonus and drain no energy to fire. When out of supply range, they fire normally and draw energy. When e-stalling, they fire more slowly. Weapon cost is default of 1, but can be defined in the FBI as customparam weaponcost.

Added groundflash CEGs to tracers - basically means that they have fake dynamic light. Still up for review.

Added a test shell trail to the 50mm weapon - also very much up for review, just kind of a whim.