Fixed transported units losing ammo.
This was caused by engine internally incrementing reloadState every frame when a
unit is stunned, it does this to prevent weapons of stunned units from reloading.
(reloadState is the frame in which reload is finished, after all.)

The gadget compared reloadState with previous value of reloadState every 3 seconds,
so when a unit was being transported, the gadget would think the unit was
continuously firing it's weapon(s).  (well, as fast as the weapon allowed.)

Now fixed by not calling ProcessWeapons and Resupply for stunned units,
and resetting the vehicles[unitID].reloadFrame[weapNum] in gadget:UnitUnloaded.

This, of course, introduced exploit that unit didn't use any ammo when you
managed to load it in a transport before the next 3-second ProcessWeapons call.
Therefore gadget:Loaded was added, this calls ProcessWeapons a final time just
before the unit is loaded in a transport, to ensure ammo is up to date.


Also fixed bug that disabling weapons didn't work properly after 55:33.3
(= 99999 frames) gametime.  This was again because reloadState is frame in
which reload is finished, and after 99999 frames the gadget would set this
to a value in the past, effectively making the weapon ready to fire.
(until TryTarget failed and the reloadTime of 99999 was used to calculate
new reloadState.)