Autogen armordef entries with 'cp.damageGroup'
This will ultimately allow us to avoid maintaining
the side armor defs entirely, but for now it is only
implemented for ships (as they're the only unit class
using OO defs in master).

This also changes ship turrets to the 'guns' damageGrou
(previously they did not have an armordefs entry).